back/requirements.txt
streamlit==1.45.1
bcrypt==4.3.0
streamlit-authenticator==0.4.2
streamlit-cookies-manager==0.2.0
extra-streamlit-components==0.1.80

pandas==2.2
numpy==1.26.4
plotly==6.1.2
yfinance==0.2.65
requests==2.31.0
beautifulsoup4==4.12.2
html5lib==1.1
python-dateutil==2.8.2
urllib3==2.4.0   

selenium==4.33.0
webdriver-manager==4.0.2

pytest==7.4.2
sendgrid==6.11.0

# API
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.12
pydantic[email]==2.10.0

sqlalchemy

# PostgreSQL 
psycopg2-binary==2.9.11

# Injecteur de variables d'environnement .env
python-dotenv==1.2.1